requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
crispy-bootstrap5==2023.10
cycler==0.12.1
Django==5.0.2
django-crispy-forms==2.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
docopt==0.6.2
fonttools==4.49.0
GDAL @ file:///C:/xampp/htdocs/geoportal/geoportal/data/GDAL-3.4.3-cp311-cp311-win_amd64.whl#sha256=f78861fb5115d5c2f8cf3c52a492ff548da9e1256dc84088947379f90e77e5b6
geoserver-rest==2.6.0
idna==3.6
Js2Py==0.74
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pipwin==0.5.2
psycopg2==2.9.9
Pygments==2.17.2
pyjsparser==2.7.1
pyparsing==3.1.2
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
seaborn==0.13.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
xmltodict==0.13.0
